Fixed working with localStorage
We check if the code is running client side, which means that:
typeof window !== "undefined" - window is only defined in browser.
If we are in a browser, then we get items from localStorage and update
our pexesosSolved and mostFoundInRow variables. Once window is defined
we render our newly fetched statistics. Also there is reactive if that
updates the web if either pexesosSolved or mostFoundInRow is changed
and the window is defined.