Remake user statistics with stores
In stores.js I created a new function that returns methods to interact
with a store that handles all the user statistics. We can get those
methods from other files and modify the store (as it is done on the
main page (+page.svelte in /routes)).

The methods were modified because they must update the local storage as
well as update the store itself.

The purpose of this modification was to make those statistics accesible
from multiple files so that we don't always have to get them from the
local storage.