Fix issue with joining lobby multiple times
When you left a lobby and joined again, you joined multiple times. This
was caused by the fact that when you left, the info was only sent to
other players, but the dictionary (lobbies) from which the state is
updated was not changed (you stayed as a player in there).